IPEC - Reduced terminal printing
IPEC now prints the first and last eigenmode energy and error. Printing every eigenmode tended to flood the terminal and push other useful information out of the user's vision.

An intelegent warning is printed if the error becomes large. Again, the full print tended to obscure any high errors in the long list.