This version includes another method of kinetic matrix calculations. The method has been used in the previous testing version of gpec (gpec_0.95) and shown the most reliable and feasible answers.
It takes place through fkg_kmats_flag, when kingridtype=0. The difference occurs by adding fluid and kinetic ABCDEH matrices and casting FKG matrices into splines, whereas the individual fluid and kinetic ABCDEH matrices are casted into splines if this flag is off.

See fourfit.f and sing.f for main changes.

The goal is to compare and understand the small differences in the results for kingridtype=0 when this flag is on and off, and ultimately to go for adaptive kinetic grids.