This is the first draft version working and showing consistent results with fkg_kmats_flag, kingridtype=0 and kingridtype=3 methods.
 - It was found that kinetic matrices were all almost identical except very minor differences for g matrix.
 - There is no simplification in g matrix.
 - There is no singularity in g matrix, but it was found g matrix has very steep variations at the core and edge, due to fluid h matrix.
 - The critical element is to spline fluid h matrix by mitigating the steep behavior, using hmats%xpower(2,:)=-1 with hmats%x0(2)=1.

Finally an oscillation in the very edge, possibly due to spurious singularity, could be eliminated without fkg_kmats_flag.

This is a draft version. In the next version, diagnostic routines will be added to provide more details, and fkg_kmats_flag will be promoted to kingridtype=0 and other methods to kingridtype=1-4.