DCON & PENTRC - Update - Big change in the DCON - PENTRC interfacing.
In editing the fourfit uses of pentrc functions to account for my recent changes to pentrc, I found that my changes to PENTRC where not well propogated to DCON. A fundemental problem was that the interfacing was done using a module that was essentially a copy-and-paste of the pentrc program, but did not need to be maintained in any way for pentrc compilation. Basically, this commit makes the pentrc program use the same module as dcon does to get all the top level pentrc commands. This will keep the dcon and pentrc programs using the same variables and commands.

 - The main sub-bullet of all this is that there were some enherent namespace conflicts between pentrc and dcon. For the most part, I avoided this through a hack-y re-naming of the pentrc variables with the use statements in dcon. But we should do better.
 -- Note: I did have to officially deprecate psiout and change the atol and rtol inputs in pentrc. **THIS WILL BREAK OLD NAMELISTS**