LSODE - Bug Fix - Fixed bug in memory allocation in two subroutines.
I assume that the compilers we use were treating it appropriately when allowed to run without a high level of debugging flags, but I think this is good to fix along with the other fixes in this patch.

I have confirmed it does not modify the dcon result in the DIII-D ideal example.