This version includes a minor but important change to make ideal version of gpec almost identical to ipec. The issue was the different ways to fit the splines when calculating fundamental matrices. See the fourfit.f for the changes.
The power extraction for hmat and gmat was essential to stabilize the kinetic solutions. It is not yet clear if this power extraction can improve ipec solutions and so is deactivated unless kin_flag is on.

The consistency was checked just for singcoup and singfld. Next the ideal examples in docs folder must be run to check if other results including ntv can also be reproduced.