WIP attempt to clean up excessive global variables.
The idea it that the external spline evaluations simply need to modify local arrays within the torque tpsi function - they do not need to modify global arrays that are used anywhere else. That is, things that want spline evaluated values outside of this function do not assume they were set inside of the function.

The first step was to get ride of all the spline specific indexing: the splines are pretty much all on the same equilibrium grid so once one finds the index of the psi crossing, we can re-use that for all the others.

The next step was to remove all the cross referenced global *_f and *_f1 type arrays and simply define them with set dimensions in the tpsi function.

The result is that the code compiles, and runs fine on one thread. On more than one, it spits out issues with finding the magnetic well, which tells me something ins off in the very first external spline evaluations (eqfun and rzphi around line 312 in torque.f90). What OMP twist am I missing?