Fixed intel compiler bugs and write formatting.
With certain compiler settings ('-C'), the intel compilers may mix up the order of .OR. statements, causing segmentation faults in the ode where some .OR.'s where assumed to be left-to-right.

I also cleaned up the formatting of some of the singular surface finding and reporting while debugging this issue (it was kinsing that was segfaulting).