INSTALL - Update - Updates default FFLAGS for official intel 2011 compiler support.
Note, I've added an initial attempt to support 2015-17 intel compilers on portal. The trick here is that Prentice recently split up the netcdf modules for C and Fortran. Right now, I can get it to compile and run ideal runs but the executable segfaults when entering the openmp part of the kinetic calculation. Thus, I kept the default flags consistent with intel 11 (e.g. -vec_report0)
Finally, this commit cleans up various un-formatted outputs to be consistent with intel formatting.