GPEC - New Output - Adds unitless Delta to the resonant coupling metrics
Delta is the published resonant coupling metric [Park, Phys. Plasmas 2007].
This commit aggregates the singcoup matrices, greatly simplifying the singcoup routine.
The commit also adds the Delta metric to singfld, although many opportunities to simplify that
routine exist and it would be nice to combine them to prevent code duplication.

This commit fixes a bug in the singcoup islandwidth calculation that is a great example of
why the singcoup and singfld subroutines need to be merged to guarantee they calculate the same thing.