GPEC - WIP - Changes "delta-prime" label to "Delta"
The classic Delta' definitions are like our jump in the derivative
Delta, but divided by the vacuum resonant field.
Right now we do not have code calculating the mutual inductance
between the control and rational surface, which we would need
to calculate the coupling matrix between Phi_x and Phi_res,vac or
the related delta-prime matrix.