VACUUM - Compiler Issue - Initializes global r variable used in io
This was motivated by the following error:
forrtl: severe (193): Run-Time Check Failure. The variable 'cardmo_$R' is being used without being defined
Which arose when debugging why a static compile of the current develop branch segfaulted when copied to and run on iris.
I believe this is related to (or exactly) the reason we segfault when compiled on more modern intel compilers (2015/2017).
It implies that our current intel 2011 compiles are ticking time bombs - assuming initializations that may or may not be there.
Note this commit fixes the known error (making r an explicit global variable).
However, ALL implicit statements are suspect.
Ideally, we should make all the vacuum code subroutines explicit in their initializations.
I tried to start that though, and it will take forever to get all the variables right.
I might try a medium step and at least make sure the missing implicit variables are integers.