PENTRC, DCON & GPEC - Update - Updates the version based on the compile-time git commit.
The main thing here is that the makefiles were updated to use a fancy representation of the latest git commit. The motivation is being able to help users debug and/or re-run old cases cleanly. Some details are,
- The version was moved out of equil globals so that not *everything* (orbit, etc) needs to be compiled when the git commit changes.
- The DCON and PENTRC versions are synced
- The GPEC version is distinct. This is so if you edit something in gpout and commit it, you can compile GPEC without having to go all the way back and compile PENTRC + DCON.

A final note is that I added some missing mod dependencies to the makefiles (especially PENTRC and GPEC), which should do a better job of truthfully catching when these need to be recompiled due to low-level changes in equil etc.