DCON - Makefile - Greatly simplifies logic to keep version in sync with pentrc
This comes at the cost of it doing the linking step for DCON every time, but that is relatively fast. I think it is worth it to ensure a) they are in sync and b) we don't unnecessarily compile all of DCON do to some weird hole in the previous complicated logic.