DCON - New Feature - Builds on edge pressure and current suppression
The DCON program re-calculates the equilibrium if the qhigh flag causes a new truncation, thus
ensuring the equil spline packing is tightest at the actual end of integration and enabling
sharp tanh suppression there.

Additional variables in dcon control the suppression of pressure and current in the fundamental
matrix splines, separate from the equil splines. This is just testing different techniques. Only one (if any) should be kept in the final version.