STRIDE - ENHANCEMENT - Adds threads/intervals/splines to stride namelist
The threads and intervals can no be specified in the namelist just like
**EVERY OTHER** controlling parameter in the code. I left backwards
compatibility, so you can override the namelist via the commandline arguments.

This commit also restores the initial STRIDE spline behavior control.
Note, as far as I could see the "new spline" was just the existing
"not-knot" spline boundary. I made this more explicit, and also added
control of the truly different "classic" spline extrapolation available
in the GPEC equil library.