DCON - ENHANCEMENT - Finishes peak_flag work
new peak_flag logic checks the dW every step after the last rational. If it went down, it stops the ODE (assumes it's reached the maximum for that rational).

I had made a spline of the vacuum matrices to save time, which assumes they are pretty smooth in the last little bit of the ODE. Seems safe (will post proof in PR).