Fixes issue with recursive directory references in superlu
I had changed the hardcoded path in make.inc to be a relative path. Turns out that relative path only worked because some makes had already happened.

Starting from scratch, it was inf looping.

Now, the remakeSuperLU.sh bash script freezes the path in after unzipping the tarball. This assumes the bash script is called in the right directory, but it works well when you run install/makefile... so we will go with this for now.