GPEC - Bug Fix - Fixes netcdf output of vsbrzphi
There was a bug in that the vacuum (r,z,phi) fields from single singular
surfaces was not being recorded as complex. They were also overwriting
each other.

Now both the netcdf and ascii record 3D fields, where one of the
dimensions is the index of the singular surface sourcing the field.
This reduces the ascii output clutter as well.