PENTRC - WIP - Initial implementation of new spline root finding
The goal is to make more robust bounce integrals
- It was suspected that irregularities in these may have been
responsible for a flood of LSODE integration warnings
- The bounce points are well identified
- The bmin and bmax are more accurately determined to better set the lambda range

This commit contains lots of debugging print statements.
It is a WIP because it is segfaulting with a "invalid memory reference"
error on my laptop (gfortran). I want to pull it to portal and see
if it happens there too. Maybe I can find it with a debugger.