PENTRC - WIP ENHANCEMENT - Attempts to parallelize pentrc bounce harmonics
This is a work in progress.
It adds a new argument (openmp_threads) to PENTRC that could be used to
speed it up. However, the computational time is already ok. The real
purpose here is to try and find the bug in the lsode integrations causing
different behavior when compiled with OMPFLAG set or not set.

The hypothesis was that while the parallelization was designed and
checked for DCON, something may have been glossed over in the pentrc
program. The idea was then to simply copy any DCON OMP behavior to PENTRC
to make sure they were behaving similarly. This included
1. Setting the number of OMP threads (even if just to 1)
2. Declaring a parallel section with the same defaults, copyins, etc.

That is pretty much it. Unfortunately, this did not solve the issue.
Still, I think it is a good step to keep.