GPEC - ENHANCEMENT - Enables GPEC when DCON used peak_flag
There was an old test demanding psilim and the end of the ODE matched.
This was a good check, since psilim is used in GPEC and would have
differed from the end of the profiles.

Instead of stopping the program, the dcon interface now prints a warning
and forces psilim and qlim to match the end of the eigenfunctions.
This enables GPEC when the ODE was cut short by the peak_flag.