DCON - Bug Fix - Makes eqfun available for inverse equilibria
This raised an error due to the more rigorous bicude_eval
checks introduced by General Fusion. I think before eqfun
was not being allocated but was somehow skating through
make_metric calls that evaluated it. With the new check,
ldp_i equilibria raised errors and could not be run.
This fixes that.