STRIDE - BUG FIX & ENHANCEMENT - Fixes netcdf output of Delta_prime
STRIDE's internal delta matrix is actually a 2Nx2N matrix of
left and right coupling coefficients.
This commit fixes a bug in which this was reformed into NxN when
being written to netcdf.

It then goes one step further and calculates & outputs the submatrices
A', B', Gamma', and Delta' described in [1]. The last of these is the NxN
matrix associated with tearing stability.

Note, this seems to be a factor of 2 off from the RDCON convention.

﻿[1] D. P. Brennan and L. E. Sugiyama, “Tearing mode stability in a low-beta plasma with sawteeth,” Physics of Plasmas, vol. 13, no. 5, p. 052515, May 2006.

closes #103