DCON - Enhancement - Improves formatting of kinetic singularity search
Note, this also attempts improvements of the newton iteration
for finding exact singularity locations. HOWEVER, I have determined that
the very use of a newton scheme is invalid here. The |F| function
is concave (looks like downward spikes), not smooth, and never exactly
zero. It cannot be approximated as a quadratic near the zero and thus
Netwon's method is not valid. If you draw some simple cartoon examples
you will quickly see that the method shoots to infinity. This is why
GPEC always stopped at the maximum iterations and printed a message
about large errors.

I have not done away with this and not replaced it because in practice
when the kinetic matrices are working we pretty much never have
singularities. So a hotfix is not necessary at this time.