PENTRC - WIP - Replaces LSODE common blocks with module variables, which are threadprivate. I think this is the modern "good practice" and it also makes tracing/handling the threadprivate variables much more clear (all the bringing out of dummy common blocks was confusing!).
This commit also adds the eqfun and rzphi splines to the threadprivate
list - just like sq, kin, etc. This seems safest.

Finally, it removes the last non-external equil spline evaluations from
pentrc. These were in a section that is pretty much never used in kinetic
DCON, but it is good to be thorough just in case this section ends up in
a parallel tpsi call at some point.

Unfortunately, all this does not fix the bug for the hotfix.
The parallel kinetic matrix calculations are still faulty - as evidenced
by the many kinetic singular surfaces still found in DCON.
See next commit for the final fix.