DCON - BUG FIX - Replaces OpenMP reduction with manual summation in a omp critical block.
The internet claims the reduction lines were valid for fortran
and the critical block is a slower implementation. However, it
definitely fails on portal when using,
intel/2018.u1 openmp/3.0.0 acml/5.3.1/ifort64

I wonder if we need to use acml/5.3.1/ifort64_mp??
This must have worked at some point. I think @aglasser17 might
have used a different intel version? It must have been older though,
so it is odd that this broke now.

Some interesting tests:
- I made a test variable that just copied the ktmat lines, except
  it simply did test = test + 1 inside the l loop.
- This worked for single numbers
- This worked for integer arrays
- It failed for real (either r8 or unspecified) or complex arrays (as ktmat fails)
So it does seem like some very subtle issue. Based on this subtlety,
I have determined the critical block is the safest fix. If I made some
fix to the reduction on portal it might not work on iris, ukstar, etc.