This is to test a mutual inductance between arbitrary two surfaces when there is no leak of flux, i.e. M=(L1#L2)^0.5 although the operations should be carefully made without hermiticity.
In reality there is always a leak, which will make mutual inductances much smaller than this ideal coupling (=1 in scalar). Fourier space needs to be restored back to a configurational space in an external shell surface as well. Next we actually utilize VACUUM routines to obtain surface currents with a shell surface in grr matrices.