INSTALL - ENHANCEMENT - Changes peak_flag to psiedge interface
I found that specifying the number of q windows to look back from the
psihigh (after sas_flag) was too sensitive because in low q95 it
could result in deep truncation while in high q95 it might not look
at enough windows.

Now, the user specifies something very intuitive: psiedge.
DCON will look at dW between this and psihigh to find the maximum dW
point for consistent truncation. This should be pretty stable.
Both low q and high q shots can use a logical edge window near the
bottom of the pedestal (if it exists).
Something like, psiedege~0.98 and psihigh~0.999.

I think this is the most intuitive for physicists