DCON - BUG FIX - Fixes bug that caused bad psi values in free_wvmats
This bug was detailed in #126. The fix is really only the change
psii = psiedge
in line 756 of fre_wvmats of free.f. This makes the finding of psi
on a regular grid of q more robust.

Also included in this commit, is the recording of the complex dW_edge.
This simply involved converting the variables from real to complex.
The calculation did not change at all (previously, the result was simply
being simplified using REAL).