Main change for this commit is to add a scan of initial guess for inboard and outboard separatrix positions. Hopefully this will make dcon run for direct equilibria more inclusive.
Another change sneaked in this version is on the pentrc ascii output files which now write actual torque densities, rather than dT/dpsi. The change is simple and can be (manually) erased if not agreed. More developments along this line (on pentrc ascii) will be posted as a pull request.