STRIDE - Bug Fix - Forces reforming of equil splines with modified bounds
STRIDE cuts the integration into intervals based on the singular surfaces
it uses the number of singular surface a lot and I couldn't trace it all.
Thus, the safest thing is just to restart everything wit the new psilow and/or
psihigh from the qlow, qhigh, sas_flag, etc.

This was motivated by cases with low q0 and reverse shear on axis. STRIDE
would fail, saying the direction of integration was wrong. I believe this was
due to a mismatch between the ODE axisPsi and the intervals when using qlow=1.02.
I also realized the old qlow calculation's initial psi guess could be tricked
by reverse shear cases with two q=1 surfaces in them. I added the improved
initial guess to DCON, since this would apply there as well.