GPEC - WIP - Adds Jbgradpsi_pest to netcdf output
Previously, relied on jac_out to get this. the vacuum field explicitly included it like the asciis though.
Adding it to the total field profiles required some logic to make sure the m_pert dimension is properly
shared between the 2 subroutines, which is probably why I was too lazy to implement this earlier. But now
I needed it to test the delta_RMP by replicating figures in Callen's report so I took the time to do it right.