GPEC - Minor - Cleans up terminal output
Now that detla_rmp requires using pentrc modules, we can use the pentrc progressbar function to report progress in the major loops.
I also continued to make more outputs into tables that are easier to read and easier to copy into tables. Specifically, I
made the surface inductances at rational surfaces a table.Note (@parkjk), I changed the convention from the average magnitude on each surface
to the magnitude of the resonant vector (sqrt of the sum of squares). I find this more consistent with the coupling matrix manipulations we do.