DCON - Work in progress - More work to check out kinetic resonances (not necessarily "singularities")
Here, we reduce the tolerances for marking a surface as a kinetic resonant surface. Previous tolerances
found very few "singularities" in practice, which is true and the whole point of including kinetic damping terms.
But now we want to diagnose the kinetic resonances even if they are not quite singular. So we have to reduce
the tolerances to mark them as special surfaces in preparation for GPEC.

Note, I found these uncommitted changes on portal after a long long time. I make no promises that they are complete. More work is needed.