DCON/STRIDE/COIL/PENTRC - ENHANCEMENT -Defaults the parallel threading to use OMP_NUM_THREADS env variable
This also deprecates the parallel_threads and openmp_threads namelist variables in favor of the more clear
dcon_kin_threads and pentrc_threads variables. This will help us as developers keep track of which settings
are where. It will also be more clear which variables do what for users. There is a soft fail for now,
so old runs can be re-run without it hard failing.

This should streamline the coordination of parallel options (i.e. default is that all codes use the same level
of parallelization).